requirements.txt
python-docx==0.8.6
lxml==4.1.1
yara-python==3.7.0
